support systems with long PIDs

+++ common/pid.h
@@ -29,14 +29,14 @@
 struct process_id {
   unsigned ip;
   short port;
-  unsigned short pid;
+  unsigned int pid;
   int utime;
 };
 
 struct process_id_ext {
   unsigned ip;
   short port;
-  unsigned short pid;
+  unsigned int pid;
   int utime;
   int actor_id;
 };
